Add back sharedConfigTypes since imports don't work otherwise for some reason
