+++ client/src/typings/redux.typings.ts
@@ -5,7 +5,7 @@ import { GroupMember } from "shared/typings/models/groups";
 import { store, combinedReducer } from "client/utils/store";
 import { UserGames, UserGroup } from "shared/typings/models/user";
 import { SignupQuestion } from "shared/typings/models/settings";
-import { LoginProvider, SignupStrategy } from "shared/config/sharedConfig";
+import { LoginProvider, SignupStrategy } from "shared/config/sharedConfigTypes";
 import { BackendErrorType } from "client/components/ErrorBar";
 import { SignupMessage } from "shared/typings/models/signupMessage";
 import { EventLogItem } from "shared/typings/models/eventLog";
