+++ client/src/views/admin/adminThunks.ts
@@ -18,7 +18,7 @@ import {
   submitSetLoginProviderAsync,
 } from "client/views/admin/adminSlice";
 import { SignupQuestion } from "shared/typings/models/settings";
-import { LoginProvider, SignupStrategy } from "shared/config/sharedConfig";
+import { LoginProvider, SignupStrategy } from "shared/config/sharedConfigTypes";
 import { getSignupMessages } from "client/services/userServices";
 import { getSentryTest } from "client/views/admin/adminService";
 import { postPlayerAssignment } from "client/services/assignmentServices";
