+++ client/src/views/all-games/allGamesUtils.ts
@@ -1,4 +1,4 @@
-import { SignupStrategy } from "shared/config/sharedConfig";
+import { SignupStrategy } from "shared/config/sharedConfigTypes";
 import { Game, ProgramType } from "shared/typings/models/game";
 
 export const getTimeslotSignupStrategy = (
