+++ client/src/views/all-games/components/GameListTitle.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from "react-i18next";
 import _ from "lodash";
 import { getTime, getWeekdayAndTime } from "client/utils/timeFormatter";
 import { SelectedGame } from "shared/typings/models/user";
-import { SignupStrategy } from "shared/config/sharedConfig";
+import { SignupStrategy } from "shared/config/sharedConfigTypes";
 import { MOBILE_MARGIN } from "client/globalStyle";
 import { getSharedConfig } from "shared/config/sharedConfig";
 import {
