+++ server/src/features/player-assignment/munkres/munkresAssignPlayers.ts
@@ -24,7 +24,7 @@ import {
   unwrapResult,
 } from "shared/utils/result";
 import { AssignmentError } from "shared/typings/api/errors";
-import { AssignmentStrategy } from "shared/config/sharedConfig";
+import { AssignmentStrategy } from "shared/config/sharedConfigTypes";
 
 export const munkresAssignPlayers = (
   players: readonly User[],
