+++ server/src/features/player-assignment/padg/padgAssignPlayers.ts
@@ -17,7 +17,7 @@ import {
   unwrapResult,
 } from "shared/utils/result";
 import { AssignmentError } from "shared/typings/api/errors";
-import { AssignmentStrategy } from "shared/config/sharedConfig";
+import { AssignmentStrategy } from "shared/config/sharedConfigTypes";
 
 export const padgAssignPlayers = (
   players: readonly User[],
