+++ server/src/features/settings/settingsController.test.ts
@@ -14,7 +14,7 @@ import { faker } from "@faker-js/faker";
 import { ApiEndpoint } from "shared/constants/apiEndpoints";
 import { UserGroup } from "shared/typings/models/user";
 import { getJWT } from "server/utils/jwt";
-import { LoginProvider, SignupStrategy } from "shared/config/sharedConfig";
+import { LoginProvider, SignupStrategy } from "shared/config/sharedConfigTypes";
 import {
   Settings,
   SignupQuestion,
