+++ shared/config/sharedConfigTypes.ts
@@ -1,0 +1,24 @@
+export enum SignupStrategy {
+  DIRECT = "direct",
+  ALGORITHM = "algorithm",
+  ALGORITHM_AND_DIRECT = "algorithm+direct",
+}
+
+export enum AssignmentStrategy {
+  MUNKRES = "munkres",
+  GROUP = "group",
+  PADG = "padg",
+  RANDOM = "random",
+  GROUP_PADG = "group+padg",
+  RANDOM_PADG = "random+padg",
+}
+
+export enum LoginProvider {
+  LOCAL = "local",
+  KOMPASSI = "kompassi",
+}
+
+export enum ConventionName {
+  ROPECON = "Ropecon",
+  TRACON_HITPOINT = "Tracon Hitpoint",
+}
